Make faster-whisper code consistent with prev changes
